Refactored tests to separate DB setup
